move voting into scoring module
