+++ app/repositories/ranking.ts
@@ -2,6 +2,8 @@ import * as Immutable from 'immutable'
 import { type Transaction, sql } from 'kysely'
 import { MAX_POSTS_PER_PAGE } from '#app/constants.ts'
 import { getEffect } from '#app/modules/scoring/effect-repository.ts'
+import { getUserVotes } from '#app/modules/scoring/vote-repository.ts'
+import { defaultVoteState } from '#app/modules/scoring/vote-service.ts'
 import { getArtefact } from '#app/repositories/artefact.ts'
 import {
 	type Effect,
@@ -25,7 +27,6 @@ import {
 	getFallacies,
 } from './post.ts'
 import { getQuote } from './quote.ts'
-import { defaultVoteState, getUserVotes } from './vote.ts'
 
 export function toImmutableReplyTree(replyTree: ReplyTree): ImmutableReplyTree {
 	return {
